+++ lib/public/views/Runs/Overview/RunsWithQcModel.js
@@ -12,6 +12,7 @@
  */
 
 import { NumericalComparisonFilterModel } from '../../../components/Filters/common/filters/NumericalComparisonFilterModel.js';
+import { RunDetectorsSelectionModel } from '../RunDetectorsSelectionModel.js';
 import { RunsOverviewModel } from './RunsOverviewModel.js';
 
 /**
@@ -27,11 +28,53 @@ export class RunsWithQcModel extends RunsOverviewModel {
     constructor(model) {
         super(model);
 
+        this._mcReproducibleAsNotBad = false;
+
+        this._runDetectorsSelectionModel = new RunDetectorsSelectionModel();
+        this._runDetectorsSelectionModel.bubbleTo(this);
+
         this.patchDisplayOptions({
             horizontalScrollEnabled: true,
             verticalScrollEnabled: true,
             freezeFirstColumn: true,
         });
+    }
+
+    /**
+     * Set mcReproducibleAsNotBad
+     *
+     * @param {boolean} mcReproducibleAsNotBad new value
+     * @return {void}
+     */
+    setMcReproducibleAsNotBad(mcReproducibleAsNotBad) {
+        this._mcReproducibleAsNotBad = mcReproducibleAsNotBad;
+        this.load();
+    }
+
+    /**
+     * Get mcReproducibleAsNotBad
+     *
+     * @return {boolean} mcReproducibleAsNotBad
+     */
+    get mcReproducibleAsNotBad() {
+        return this._mcReproducibleAsNotBad;
+    }
+
+    /**
+     * Returns the run detectors selection model
+     *
+     * @return {RunDetectorsSelectionModel} selection model
+     */
+    get runDetectorsSelectionModel() {
+        return this._runDetectorsSelectionModel;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    async load() {
+        this._runDetectorsSelectionModel.reset();
+        super.load();
     }
 
     /**
