+++ lib/public/views/Runs/RunPerDataPass/RunsPerDataPassOverviewModel.js
@@ -20,7 +20,6 @@ import { jsonPatch } from '../../../utilities/fetch/jsonPatch.js';
 import { jsonPut } from '../../../utilities/fetch/jsonPut.js';
 import { SkimmingStage } from '../../../domain/enums/SkimmingStage.js';
 import { NumericalComparisonFilterModel } from '../../../components/Filters/common/filters/NumericalComparisonFilterModel.js';
-import { RunDetectorsSelectionModel } from '../RunDetectorsSelectionModel.js';
 import { jsonFetch } from '../../../utilities/fetch/jsonFetch.js';
 
 /**
@@ -44,8 +43,6 @@ export class RunsPerDataPassOverviewModel extends FixedPdpBeamTypeRunsOverviewMo
         this._gaqSummary$ = new ObservableData(RemoteData.notAsked());
         this._gaqSummary$.bubbleTo(this);
 
-        this._mcReproducibleAsNotBad = false;
-
         this._markAsSkimmableRequestResult$ = new ObservableData(RemoteData.notAsked());
         this._markAsSkimmableRequestResult$.bubbleTo(this);
 
@@ -56,9 +53,6 @@ export class RunsPerDataPassOverviewModel extends FixedPdpBeamTypeRunsOverviewMo
             scale: 0.01,
             integer: false,
         }));
-
-        this._runDetectorsSelectionModel = new RunDetectorsSelectionModel();
-        this._runDetectorsSelectionModel.bubbleTo(this);
 
         this._freezeOrUnfreezeActionState$ = new ObservableData(RemoteData.notAsked());
         this._freezeOrUnfreezeActionState$.bubbleTo(this);
@@ -109,8 +103,6 @@ export class RunsPerDataPassOverviewModel extends FixedPdpBeamTypeRunsOverviewMo
         if (!this._dataPassId) {
             return;
         }
-
-        this._runDetectorsSelectionModel.reset();
 
         this._fetchQcSummary();
         this._fetchGaqSummary();
@@ -306,35 +298,6 @@ export class RunsPerDataPassOverviewModel extends FixedPdpBeamTypeRunsOverviewMo
      */
     get dataPassId() {
         return this._dataPassId;
-    }
-
-    /**
-     * Set mcReproducibleAsNotBad
-     *
-     * @param {boolean} mcReproducibleAsNotBad new value
-     * @return {void}
-     */
-    setMcReproducibleAsNotBad(mcReproducibleAsNotBad) {
-        this._mcReproducibleAsNotBad = mcReproducibleAsNotBad;
-        this.load();
-    }
-
-    /**
-     * Get mcReproducibleAsNotBad
-     *
-     * @return {boolean} mcReproducibleAsNotBad
-     */
-    get mcReproducibleAsNotBad() {
-        return this._mcReproducibleAsNotBad;
-    }
-
-    /**
-     * Returns the run detectors selection model
-     *
-     * @return {RunDetectorsSelectionModel} selection model
-     */
-    get runDetectorsSelectionModel() {
-        return this._runDetectorsSelectionModel;
     }
 
     /**
