+++ lib/public/views/Runs/RunsPerSimulationPass/RunsPerSimulationPassOverviewModel.js
@@ -15,7 +15,6 @@ import { RunsWithQcModel } from '../Overview/RunsWithQcModel.js';
 import { ObservableData } from '../../../utilities/ObservableData.js';
 import { getRemoteData } from '../../../utilities/fetch/getRemoteData.js';
 import { detectorsProvider } from '../../../services/detectors/detectorsProvider.js';
-import { RunDetectorsSelectionModel } from '../RunDetectorsSelectionModel.js';
 
 /**
  * Runs Per Simulation Pass overview model
@@ -29,13 +28,12 @@ export class RunsPerSimulationPassOverviewModel extends RunsWithQcModel {
         super(model);
         this._detectors$ = detectorsProvider.qc$;
         this._detectors$.bubbleTo(this);
+        this.registerDetectorsNotBadFractionFilterModels(this._detectors$);
+
         this._simulationPass$ = new ObservableData(RemoteData.notAsked());
         this._simulationPass$.bubbleTo(this);
         this._qcSummary$ = new ObservableData(RemoteData.NotAsked());
         this._qcSummary$.bubbleTo(this);
-
-        this._runDetectorsSelectionModel = new RunDetectorsSelectionModel();
-        this._runDetectorsSelectionModel.bubbleTo(this);
     }
 
     /**
@@ -70,8 +68,6 @@ export class RunsPerSimulationPassOverviewModel extends RunsWithQcModel {
      * @inheritdoc
      */
     async load() {
-        this._runDetectorsSelectionModel.reset();
-
         this._fetchQcSummary();
         this._fetchSimulationPass();
         super.load();
@@ -128,14 +124,5 @@ export class RunsPerSimulationPassOverviewModel extends RunsWithQcModel {
      */
     get qcSummary() {
         return this._qcSummary$.getCurrent();
-    }
-
-    /**
-     * Returns the run detectors selection model
-     *
-     * @return {RunDetectorsSelectionModel} selection model
-     */
-    get runDetectorsSelectionModel() {
-        return this._runDetectorsSelectionModel;
     }
 }
