+++ lib/public/views/Runs/RunsPerSimulationPass/RunsPerSimulationPassOverviewPage.js
@@ -92,7 +92,7 @@ export const RunsPerSimulationPassOverviewPage = ({
 
             return [
                 h('.flex-row.justify-between.items-center.g2', [
-                    filtersPanelPopover(perSimulationPassOverviewModel.filteringModel, activeColumns, { profile: 'runsPerSimulationPass' }),
+                    filtersPanelPopover(perSimulationPassOverviewModel, activeColumns, { profile: 'runsPerSimulationPass' }),
                     h('.pl2#runOverviewFilter', runNumbersFilter(perSimulationPassOverviewModel.filteringModel.get('runNumbers'))),
                     h(
                         '.flex-row.g1.items-center',
